Optimized Book Controller Code for best practices created Customer controller - pending associated view files Written Test cases for customer controller

+++ app/Http/Controllers/BookController.php
@@ -14,7 +14,7 @@ use Illuminate\Support\Facades\Storage;
 class BookController extends Controller
 {
     /**
-     * Display a listing of the resource.
+     * Display a listing of the Books.
      *
      * @return Application|Factory|View
      */
@@ -24,9 +24,9 @@ class BookController extends Controller
     }
 
     /**
-     * Show the form for creating a new resource.
+     * Show the form for creating a new book.
      *
-     * @return \Illuminate\Http\Response
+     * @return Application|Factory|View
      */
     public function create()
     {
@@ -36,46 +36,16 @@ class BookController extends Controller
     /**
      * Store a newly created valid book in books table.
      *
-     * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\Response
+     * @param StoreBookRequest $request
+     * @return RedirectResponse
      */
     public function store(StoreBookRequest $request)
     {
         $input = $request->validated();
-
-        if ($input['mode'] === 'online') {
-            $book = Book::create([
-                'name' => $input['name'] ,
-                'author' => $input['author'] ,
-                'price' => $input['price'] ,
-                'version' => $input['version'] ,
-                'mode' => $input['mode'] ,
-                'is_download_allowed' => $input['is_download_allowed'] ,
-                'book_path' => $input['book']->store('books') ,
-                'image' => $input['image']->store('data/books/front-covers' , ['disk' => 'public'])
-            ]);
-        } else {
-            $book = Book::create([
-                'name' => $input['name'] ,
-                'author' => $input['author'] ,
-                'price' => $input['price'] ,
-                'version' => $input['version'] ,
-                'mode' => $input['mode'] ,
-                'image' => $input['image']->store('data/books/front-covers' , ['disk' => 'public'])]);
-        }
+        $input = $this->storeUploadedFilesByArray($input);
+        Book::create($input);
 
         return to_route('books.index');
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param int $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
     }
 
     /**
@@ -104,17 +74,7 @@ class BookController extends Controller
         $book->price = $request->price;
         $book->version = $request->version;
         $book->mode = $request->mode;
-
-        if (isset($request->image)) {
-            $book->image = $request->image->store('data/books/front-covers' , ['disk' => 'public']);
-        }
-
-        if ($request->file('book_file') !== null && $request->mode === 'online') {
-
-            $book->book_path = $request->file('book_file')->store('books');
-            $book->is_download_allowed = $request->is_download_allowed;
-        }
-
+        $this->storeUploadedFiles($request , $book);
         $book->save();
 
         return to_route('books.index');
@@ -129,16 +89,53 @@ class BookController extends Controller
      */
     public function destroy(Book $book): RedirectResponse
     {
-        //Delete a Book Front Image from Storage if Present
-        if ($book->image)
-            Storage::disk('public')->delete($book->image);
-
-        //Delete a Book PDF File from Storage if present
-        if ($book->book_path)
-            Storage::delete($book->book_path);
-
+        $this->deleteBookFiles($book);
         $book->delete();
 
         return to_route('books.index');
+    }
+
+    /**
+     * Store the uploaded image and pdf file
+     * @param array $input
+     * @return array
+     */
+    public function storeUploadedFilesByArray(array $input): array
+    {
+        $input['book_path'] = array_key_exists('book' , $input) ? $input['book']->store(config('filesystems.book_pdf_files')) : null;
+        $input['image'] = $input['image']->store(config('filesystems.book_front_covers') , ['disk' => 'public']);
+        return $input;
+    }
+
+    /**
+     * @param UpdateBookRequest $request
+     * @param Book $book
+     * @return void
+     */
+    public function storeUploadedFiles(UpdateBookRequest $request , Book $book): void
+    {
+        if (isset($request->image)) {
+            $book->image = $request->image->store(config('filesystems.book_front_covers') , ['disk' => 'public']);
+        }
+
+        if ($request->file('book_file') !== null && $request->mode === 'online') {
+            $book->book_path = $request->file('book_file')->store(config('filesystems.book_pdf_files'));
+            $book->is_download_allowed = $request->is_download_allowed;
+        }
+    }
+
+    /**
+     *
+     * Delete Book PDF File and Front Image
+     * @param Book $book
+     * @return void
+     */
+    public function deleteBookFiles(Book $book): void
+    {
+        //Delete a Book Front Image from Storage if Present
+        if ($book->image) Storage::disk('public')->delete($book->image);
+
+        //Delete a Book PDF File from Storage if present
+        if ($book->book_path) Storage::delete($book->book_path);
     }
 }
